Create a new pull request by comparing

+++ test/automation/src/code.ts
@@ -63,7 +63,7 @@ function getBuildOutPath(root: string): string {
 	}
 }
 
-async function connect(connectDriver: typeof connectElectronDriver, child: cp.ChildProcess | undefined, outPath: string, handlePath: string, logger: Logger): Promise<Code> {
+async function connect(connectDriver: typeof connectElectronDriver | typeof connectPlaywrightDriver, child: cp.ChildProcess | undefined, outPath: string, handlePath: string, logger: Logger): Promise<Code> {
 	let errCount = 0;
 
 	while (true) {
@@ -79,7 +79,7 @@ async function connect(connectDriver: typeof connectElectronDriver, child: cp.Ch
 			}
 
 			// retry
-			await new Promise(c => setTimeout(c, 100));
+			await new Promise(resolve => setTimeout(resolve, 100));
 		}
 	}
 }
@@ -116,14 +116,12 @@ export async function spawn(options: SpawnOptions): Promise<Code> {
 	const handle = await createDriverHandle();
 
 	let child: cp.ChildProcess | undefined;
-	let connectDriver: typeof connectElectronDriver;
 
 	copyExtension(options.extensionsPath, 'vscode-notebook-tests');
 
 	if (options.web) {
 		await launch(options.userDataDir, options.workspacePath, options.codePath, options.extensionsPath, Boolean(options.verbose));
-		connectDriver = connectPlaywrightDriver.bind(connectPlaywrightDriver, options);
-		return connect(connectDriver, child, '', handle, options.logger);
+		return connect(connectPlaywrightDriver.bind(connectPlaywrightDriver, options), child, '', handle, options.logger);
 	}
 
 	const env = { ...process.env };
@@ -199,8 +197,7 @@ export async function spawn(options: SpawnOptions): Promise<Code> {
 	child = cp.spawn(electronPath, args, spawnOptions);
 	instances.add(child);
 	child.once('exit', () => instances.delete(child!));
-	connectDriver = connectElectronDriver;
-	return connect(connectDriver, child, outPath, handle, options.logger);
+	return connect(connectElectronDriver, child, outPath, handle, options.logger);
 }
 
 async function copyExtension(extensionsPath: string, extId: string): Promise<void> {
@@ -290,34 +287,55 @@ export class Code {
 		await this.driver.dispatchKeybinding(windowId, keybinding);
 	}
 
-	async reload(): Promise<void> {
-		const windowId = await this.getActiveWindowId();
-		await this.driver.reloadWindow(windowId);
-	}
-
 	async exit(): Promise<void> {
-		const exitPromise = this.driver.exitApplication();
+		return new Promise<void>((resolve, reject) => {
+			let done = false;
 
-		// If we know the `pid`, use that to await the
-		// process to terminate (desktop).
-		const pid = this.pid;
-		if (typeof pid === 'number') {
-			await (async () => {
-				while (true) {
-					try {
-						process.kill(pid, 0); // throws an exception if the main process doesn't exist anymore.
-						await new Promise(c => setTimeout(c, 100));
-					} catch (error) {
-						return;
-					}
+			// Start the exit flow via driver
+			const exitPromise = this.driver.exitApplication().then(veto => {
+				if (veto) {
+					done = true;
+					reject(new Error('Smoke test exit call resulted in unexpected veto'));
 				}
-			})();
-		}
+			});
 
-		// Otherwise await the exit promise (web).
-		else {
-			await exitPromise;
-		}
+			// If we know the `pid` of the smoke tested application
+			// use that as way to detect the exit of the application
+			const pid = this.pid;
+			if (typeof pid === 'number') {
+				(async () => {
+					let killCounter = 0;
+					while (!done) {
+						killCounter++;
+
+						if (killCounter > 40) {
+							done = true;
+							reject(new Error('Smoke test exit call did not terminate main process after 20s, giving up'));
+						}
+
+						try {
+							process.kill(pid, 0); // throws an exception if the main process doesn't exist anymore.
+							await new Promise(resolve => setTimeout(resolve, 500));
+						} catch (error) {
+							done = true;
+							resolve();
+						}
+					}
+				})();
+			}
+
+			// Otherwise await the exit promise (web).
+			else {
+				(async () => {
+					try {
+						await exitPromise;
+						resolve();
+					} catch (error) {
+						reject(new Error(`Smoke test exit call resulted in error: ${error}`));
+					}
+				})();
+			}
+		});
 	}
 
 	async waitForTextContent(selector: string, textContent?: string, accept?: (result: string) => boolean, retryCount?: number): Promise<string> {
